Reorganized and cleaned up
