+++ TODO.md
@@ -3,3 +3,5 @@
 TODOs that emerge as I port the code.
 
 - Go through the model spec and make sure the numbers are right. Add cache pricing.
+- Make sure we are scaling temperature correctly
+- Update the way the client use the providers to match the rewritten configs
