+++ kintu/types/content.py
@@ -1,21 +1,30 @@
-from pydantic import BaseModel, ConfigDict
-from PIL import Image
+from __future__ import annotations
+
 from typing import Any
+
+from PIL import Image
+from pydantic import BaseModel, ConfigDict
+
 
 # Content types
 class Content(BaseModel):
     """Base class for all content types."""
+
     model_config = ConfigDict(arbitrary_types_allowed=True)
+
 
 class TextContent(Content):
     text: str
 
+
 class ImageContent(Content):
     image: Image.Image
+
 
 class DocumentContent(Content):
     document: bytes
     openai_filename: str | None = None
+
 
 class ThinkingContent(Content):
     thinking: str
@@ -23,8 +32,10 @@ class ThinkingContent(Content):
     # For multi-turn, we need to preserve encrypted thinking data
     encrypted_data: Any | None = None
 
+
 class AnthropicRedactedThinkingContent(Content):
     data: str
+
 
 class ToolCallContent(Content):
     tool_id: str
@@ -33,17 +44,21 @@ class ToolCallContent(Content):
 
     openai_id: str | None = None
 
+
 class ToolResultContent(Content):
     tool_id: str
     results: list[Content]  # Can contain TextContent, ImageContent, etc.
     is_error: bool = False
 
+
 # Anthropic-specific content types
 class AnthropicServerToolUse(Content):
     """Anthropic server-side tool use (web search, code execution, etc.)"""
+
     tool_id: str
     tool_name: str
     input: dict[str, Any] = {}
+
 
 class AnthropicWebSearchIndividualResult(BaseModel):
     encrypted_content: str
@@ -51,8 +66,10 @@ class AnthropicWebSearchIndividualResult(BaseModel):
     title: str
     url: str
 
+
 class AnthropicWebSearchToolResult(Content):
     """Anthropic web search tool result"""
+
     search_results: list[AnthropicWebSearchIndividualResult]
     tool_use_id: str
     error: str | None = None
@@ -60,6 +77,7 @@ class AnthropicWebSearchToolResult(Content):
 
 class AnthropicCodeInterpreterToolResult(Content):
     """Anthropic code interpreter tool result"""
+
     tool_use_id: str
     return_code: int
     stdout: str
@@ -71,28 +89,53 @@ class AnthropicCodeInterpreterToolResult(Content):
 # OpenAI-specific content types
 class OpenAIServerToolUse(Content):
     """OpenAI server-side tool use"""
+
     tool_id: str
     tool_name: str
     tool_type: str  # e.g., 'web_search_preview_2025_03_11', 'code_interpreter'
     input: dict[str, Any] = {}
 
+
 # Gemini-specific content types
 class GeminiServerToolUse(Content):
     """Gemini server-side tool use (web search, code execution, etc.)"""
+
     tool_id: str
     tool_name: str
     tool_type: str  # 'google_search', 'code_execution', etc.
     input: dict[str, Any] = {}
 
+
 class GeminiWebSearchResult(Content):
     """Gemini web search result"""
+
     tool_use_id: str
-    search_results: list[dict]  # List of search result dicts
+    search_results: list[dict[str, Any]]  # List of search result dicts
     error: str | None = None
+
 
 class GeminiCodeExecutionResult(Content):
     """Gemini code execution result"""
+
     tool_use_id: str
     outcome: str  # 'OUTCOME_OK' or error
     output: str | None = None
     error: str | None = None
+
+
+ContentTypes = (
+    TextContent
+    | ImageContent
+    | DocumentContent
+    | ThinkingContent
+    | ToolCallContent
+    | ToolResultContent
+    | AnthropicRedactedThinkingContent
+    | AnthropicServerToolUse
+    | AnthropicWebSearchToolResult
+    | AnthropicCodeInterpreterToolResult
+    | OpenAIServerToolUse
+    | GeminiServerToolUse
+    | GeminiWebSearchResult
+    | GeminiCodeExecutionResult
+)
