+++ kintu/types/errors.py
@@ -1,9 +1,16 @@
+class KintuError(Exception):
+    """Base class for all Kintu errors."""
 
-
-class MissingCredentialsError(Exception):
-    """Raised when required credentials for a LiteLLM backend are missing."""
     pass
 
-class InvalidLiteLLMModelError(Exception):
+
+class MissingCredentialsError(KintuError):
+    """Raised when credentials are missing."""
+
+    pass
+
+
+class InvalidLiteLLMModelError(KintuError):
     """Raised when a model is not a valid LiteLLM model."""
+
     pass
