+++ kintu/types/llmsdks/README.md
@@ -1,0 +1,6 @@
+# LLM SDK types
+
+Types from the underlying LLM SDKs, collected and re-exported with a `Raw*` prefix. 
+
+For convenience, but also because we wrap the SDKs while exposing the underlying inputs/outputs, we rely on
+the underlying types fairly commonly. And the SDKs do not always have well organized/easy-to-find types.
