+++ kintu/types/llmsdks/gemini.py
@@ -1,0 +1,80 @@
+__all__ = [
+    "RawGeminiGenerateContentResponse",
+    "RawGeminiContent",
+    "RawGeminiPart",
+    "RawGeminiCandidate",
+    "RawGeminiFunctionCall",
+    "RawGeminiFunctionResponse",
+    "RawGeminiExecutableCode",
+    "RawGeminiCodeExecutionResult",
+    "RawGeminiThinkingConfig",
+    "RawGeminiGoogleSearch",
+    "RawGeminiGoogleSearchRetrieval",
+    "RawGeminiBlob",
+    "RawGeminiTool",
+    "RawGeminiFunctionDeclaration",
+    "RawGeminiGenerationConfig",
+    "RawGeminiUsageMetadata",
+    "RawGeminiGenerateContentConfig",
+    "RawGeminiToolConfig",
+    "RawGeminiFunctionCallingConfig",
+]
+
+# Import Gemini types with RawGemini* naming pattern
+from google.genai.types import (
+    Blob as RawGeminiBlob,
+)
+from google.genai.types import (
+    Candidate as RawGeminiCandidate,
+)
+from google.genai.types import (
+    CodeExecutionResult as RawGeminiCodeExecutionResult,
+)
+from google.genai.types import (
+    Content as RawGeminiContent,
+)
+from google.genai.types import (
+    ExecutableCode as RawGeminiExecutableCode,
+)
+from google.genai.types import (
+    FunctionCall as RawGeminiFunctionCall,
+)
+from google.genai.types import (
+    FunctionCallingConfig as RawGeminiFunctionCallingConfig,
+)
+from google.genai.types import (
+    FunctionDeclaration as RawGeminiFunctionDeclaration,
+)
+from google.genai.types import (
+    FunctionResponse as RawGeminiFunctionResponse,
+)
+from google.genai.types import (
+    GenerateContentConfig as RawGeminiGenerateContentConfig,
+)
+from google.genai.types import (
+    GenerateContentResponse as RawGeminiGenerateContentResponse,
+)
+from google.genai.types import (
+    GenerateContentResponseUsageMetadata as RawGeminiUsageMetadata,
+)
+from google.genai.types import (
+    GenerationConfig as RawGeminiGenerationConfig,
+)
+from google.genai.types import (
+    GoogleSearch as RawGeminiGoogleSearch,
+)
+from google.genai.types import (
+    GoogleSearchRetrieval as RawGeminiGoogleSearchRetrieval,
+)
+from google.genai.types import (
+    Part as RawGeminiPart,
+)
+from google.genai.types import (
+    ThinkingConfig as RawGeminiThinkingConfig,
+)
+from google.genai.types import (
+    Tool as RawGeminiTool,
+)
+from google.genai.types import (
+    ToolConfig as RawGeminiToolConfig,
+)
