+++ kintu/types/message.py
@@ -1,6 +1,8 @@
 from pydantic import BaseModel, ConfigDict
+
 from kintu.types.content import Content
 from kintu.types.role import Role
+
 
 class Message(BaseModel):
     role: Role
