+++ kintu/types/role.py
@@ -1,5 +1,6 @@
 import enum
 
+
 class Role(str, enum.Enum):
     USER = "user"
     ASSISTANT = "assistant"
