+++ kintu/types/tool.py
@@ -1,5 +1,6 @@
 from pydantic import BaseModel
 
+
 class Tool(BaseModel):
     name: str
     description: str
