+++ kintu/utilities.py
@@ -1,9 +1,11 @@
-from PIL import Image
 import base64
 import io
+
+from PIL import Image
+
 
 def image_to_base64(image: Image.Image) -> str:
     """Convert PIL Image to base64 string."""
     buffer = io.BytesIO()
-    image.save(buffer, format='PNG')
-    return base64.b64encode(buffer.getvalue()).decode('utf-8')
+    image.save(buffer, format="PNG")
+    return base64.b64encode(buffer.getvalue()).decode("utf-8")
