+++ pyproject.toml
@@ -22,6 +22,7 @@ classifiers = [
     "Programming Language :: Python :: 3.13",
 ]
 dependencies = [
+    "google-genai>=1.29.0",
     "pillow>=11.3.0",
     "pydantic>=2.11.7",
 ]
